requirements.txt
python_nmap==0.7.1
scapy==2.6.1
tabulate==0.9.0
